// ya-ryadom-ui/src/utils/enums/views.enum.ts
export enum VIEWS {
    MY_PROFILE_VIEW = 'MY_PROFILE_VIEW',
    EVENTS_NEAR_ME_VIEW = 'PEOPLE_NEAR_ME_VIEW',
    REVIEWS_VIEW = 'REVIEWS_VIEW',
    APPLICATIONS_VIEW = 'APPLICATIONS_VIEW',
    INTRO_VIEW = 'INTO_VIEW',
    GENERAL_VIEW = 'GENERAL_VIEW',
    MY_EVENT_CREATE_VIEW = 'MY_EVENT_CREATE_VIEW',
}